ZYB-003 небольшой рефакторинг логики примеров, затащил на проект Lobmok
