+++ src/main/java/ru/ikusov/training/telegrambot/services/ExampleAnswerMessageGenerator.java
@@ -1,5 +1,7 @@
 package ru.ikusov.training.telegrambot.services;
 
+import org.telegram.telegrambots.meta.api.objects.Chat;
+import org.telegram.telegrambots.meta.api.objects.User;
 import ru.ikusov.training.telegrambot.dao.DatabaseConnector;
 import ru.ikusov.training.telegrambot.model.ChatEntity;
 import ru.ikusov.training.telegrambot.model.UserEntity;
@@ -11,25 +13,40 @@ import static ru.ikusov.training.telegrambot.MainClass.AVTFTALK_CHAT_ID;
 
 public class ExampleAnswerMessageGenerator {
     private MathAchieves achieves;
-    private UserEntity user;
+    private final String userName;
     private final int userAnswer;
     private final int rightAnswer;
     private final boolean isRight;
-    private final boolean avtftalk;
+    private final boolean isAvtftalk;
     private final int timer;
     private int exampleScore;
     private int sumScore;
 
     public ExampleAnswerMessageGenerator(DatabaseConnector databaseConnector, UserEntity user,
                                          ChatEntity chat, int userAnswer, int rightAnswer, int timer) {
-        this.avtftalk = String.valueOf(chat.getId()).equals(AVTFTALK_CHAT_ID);
+        this.isAvtftalk = String.valueOf(chat.getId()).equals(AVTFTALK_CHAT_ID);
         this.userAnswer = userAnswer;
         this.rightAnswer = rightAnswer;
         this.isRight = userAnswer==rightAnswer;
         this.timer = timer;
-        this.user = user;
-        if (isRight && avtftalk) {
+        this.userName = UserNameGetter.getUserName(user);
+        if (isRight && isAvtftalk) {
             achieves = new MathAchieves(databaseConnector, user, chat, rightAnswer, timer);
+            exampleScore = achieves.getSum();
+            sumScore = achieves.getScore();
+        }
+    }
+
+    public ExampleAnswerMessageGenerator(MathAchieves achieves, User user,
+                                         Chat chat, int userAnswer, int rightAnswer, int timer) {
+        this.isAvtftalk = String.valueOf(chat.getId()).equals(AVTFTALK_CHAT_ID);
+        this.userAnswer = userAnswer;
+        this.rightAnswer = rightAnswer;
+        this.isRight = userAnswer==rightAnswer;
+        this.timer = timer;
+        this.userName = UserNameGetter.getUserName(user);
+        if (isRight && isAvtftalk) {
+            this.achieves = achieves;
             exampleScore = achieves.getSum();
             sumScore = achieves.getScore();
         }
@@ -37,14 +54,14 @@ public class ExampleAnswerMessageGenerator {
 
     public String generate() {
         return isRight ?
-                avtftalk ?
+                isAvtftalk ?
                 generateRightAnswerMessage() :
                 generateSimpleRightAnswerMessage() :
                 generateWrongAnswerMessage();
     }
 
     private String generateWrongAnswerMessage() {
-        int dif = Math.abs(userAnswer-rightAnswer);
+        int dif = Math.abs(userAnswer - rightAnswer);
         if (dif<6) return "So close...";
         if (dif==10) return "Десяточка обиженно смотрит на тебя!";
         if (dif==100) return "Соточка обиженно смотрит на тебя!";
@@ -55,14 +72,12 @@ public class ExampleAnswerMessageGenerator {
     private String generateSimpleRightAnswerMessage() {
         return String.format(
                 MessageType.RIGHT_ANSWER_MESSAGE.getRandomMessage(),
-                String.valueOf(userAnswer),
-                UserNameGetter.getUserName(user),
+                userAnswer,
+                userName,
                 MyMath.secondsToReadableTime(timer));
     }
 
     private String generateRightAnswerMessage() {
-        String userName = UserNameGetter.getUserName(user);
-
         StringBuilder msgB = new StringBuilder();
 
         msgB.append(String.format("Победитель - %s. Так держать! ", userName));
